Task N+6: Client done. Onto ExEn.

+++ src/utility/random/random_utils.cpp
@@ -13,7 +13,7 @@ namespace rando {
         };
     }
 
-    std::chrono::milliseconds Delay(const int min, const int max) {
+    std::chrono::milliseconds Delay(const size_t min, const size_t max) {
         const std::chrono::milliseconds t_min {min}, t_max {max};
 
         std::uniform_int_distribution dist (t_min.count(), t_max.count());
